requirements.txt
langgraph
ollama
torch
transformers
peft
Pillow
fastapi
uvicorn
python-multipart
pydantic